+++ frontend/src/pages/positions/components/index.ts
@@ -1,6 +1,7 @@
 export { OverallStatsPanel } from './OverallStatsPanel'
 export { FilterPanel } from './FilterPanel'
 export { PositionCard } from './PositionCard'
+export { AIReasoningExpander } from './AIReasoningExpander'
 export { CloseConfirmModal } from './modals/CloseConfirmModal'
 export { NotesModal } from './modals/NotesModal'
 export { TradeHistoryModal } from './modals/TradeHistoryModal'